Avoid stray empty unused buffers when opening multiple files
Prior to this change, when I wanted to open a new file with its
associated files (templates, tests, etc), I would:

1. Ctrl-t to open a new tab
2. Ctrl-p to open telescope
3. Select multiple files to open
4. Ctrl-v to open the files in a split

Previously, this would leave an empty, unused buffer as the leftmost
split, which I would have to go over and close.

This change automates the closing of that unused buffer.